+++ integration-tests/library/src/main/java/com/squareup/anvil/test/Bindings.kt
@@ -2,6 +2,7 @@ package com.squareup.anvil.test
 
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
+import javax.inject.Named
 
 public interface ParentType
 
@@ -14,4 +15,5 @@ public interface MiddleType : ParentType
 public class AppBinding @Inject constructor() : MiddleType
 
 @ContributesBinding(SubScope::class)
+@Named("middle")
 public object SubcomponentBinding : MiddleType